Ints to float64

+++ calstats.go
@@ -6,14 +6,14 @@ import (
 	sc "math-skills/statcalc"
 ) 
 
-func CalculateAndPrintStatistics(data []int) {
+func CalculateAndPrintStatistics(data []float64) {
 	average := sc.CalculateAverage(data)
 	median := sc.CalculateMedian(data)
 	variance := sc.CalculateVariance(data, average)
 	stdDeviation := sc.CalculateStandardDeviation(variance)
 
-	fmt.Printf("Average: %d\n", average)
-	fmt.Printf("Median: %d\n", median)
-	fmt.Printf("Variance: %d\n", variance)
-	fmt.Printf("Standard Deviation: %d\n", stdDeviation)
+	fmt.Printf("Average: %f\n", average)
+	fmt.Printf("Median: %f\n", median)
+	fmt.Printf("Variance: %f\n", variance)
+	fmt.Printf("Standard Deviation: %f\n", stdDeviation)
 }
