+++ statcalc/deviation.go
@@ -2,6 +2,6 @@ package statcalc
 
 import "math"
 
-func CalculateStandardDeviation(variance int) int {
-	return int(math.Sqrt(float64(variance)))
+func CalculateStandardDeviation(variance float64) float64 {
+	return math.Sqrt(variance)
 }
